Restructure the files
