+++ engine/modernizers/static_site/rules/jquery_rules.py
