+++ engine/modernizers/static_site/transpilers/__init__.py
@@ -1,0 +1,10 @@
+"""
+Static Site Transpilers Module.
+
+This module contains transpilers for converting legacy static websites
+to modern frameworks like React, Next.js, and Astro.
+"""
+
+from .static_site_transpiler import StaticSiteTranspiler
+
+__all__ = ['StaticSiteTranspiler']
